update the example.
Now the Fade have 2 Version.
One way to use use the Update function.
Or Using a Coroutine.

Honestly the main goal was to avoid using coroutines...but after the issue with the android test, probably we should need both versions.
Never think about this case need to organize the scripts.